Moving message display into a separate library in preparation of unit testing.
